requirements.txt
google-generativeai>=0.3.0
click>=8.0.0
rich>=12.0.0
pyyaml>=6.0 